Improvements to code organization & api/docs
* Move supported dataset definitions into their own module
* Rename `ALL_DATASETS` to `ALL_SUPPORTED_DATASETS`
* `datasets` kwarg for `find_iceflow_data` now more directly supports users
specifying a dataset using a shortname and version, either through `Datset`
instantiation or a list of dicts with short name & version.
* Rerender notebooks
