+++ docs/getting-started.md
@@ -39,16 +39,31 @@ By default, all iceflow-supported datasets are searched. To search for a
 specific subset of iceflow-supported datasets, use the `datasets` kwarg:
 
 ```
-from nsidc.iceflow import ILATM1BDataset
+from nsidc.iceflow import Dataset
 
 
 search_results = find_iceflow_data(
-    datasets=[ILATM1BDataset(version="1")],
+    datasets=[Dataset(short_name="ILATM1B", version="1")],
     # Lower_left_lon, lower_left_lat, upper_right_lon, upper_right_lat
     bounding_box=(-103.125559, -75.180563, -102.677327, -74.798063),
     temporal=(dt.date(2009, 11, 1), dt.date(2009, 12, 31)),
 )
 ```
+
+```{include} ../supported_datasets
+
+```
+
+`iceflow` currently supports the following datasets:
+
+| Dataset                                                  | Temporal Coverage             |
+| -------------------------------------------------------- | ----------------------------- |
+| [ILATM1B v1](https://nsidc.org/data/ilatm1b/versions/1)  | 2009-03-31 through 2012-11-08 |
+| [ILATM1B v2](https://nsidc.org/data/ilatm1b/versions/2)  | 2013-03-20 through 2019-11-20 |
+| [BLATM1B v1](https://nsidc.org/data/blatm1b/versions/1)  | 1993-06-23 through 2008-10-30 |
+| [ILVIS2 v1](https://nsidc.org/data/ilvis2/versions/1)    | 2009-04-14 through 2015-10-31 |
+| [ILVIS2 v2](https://nsidc.org/data/ilvis2/versions/2)    | 2017-08-25 through 2017-09-20 |
+| [GLAH06 v034](https://nsidc.org/data/glah06/versions/34) | 2003-02-20 through 2009-10-11 |
 
 All other keyword arguments to this function (e.g,. `bounding_box`, `temporal`)
 map to [CMR](https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html)
