+++ src/nsidc/iceflow/data/__init__.py
@@ -1,16 +1,18 @@
 from __future__ import annotations
 
 from nsidc.iceflow.data.models import (
-    ALL_DATASETS,
-    BLATM1BDataset,
     Dataset,
+)
+from nsidc.iceflow.data.supported_datasets import (
+    ALL_SUPPORTED_DATASETS,
+    BLATM1BDataset,
     GLAH06Dataset,
     ILATM1BDataset,
     ILVIS2Dataset,
 )
 
 __all__ = [
-    "ALL_DATASETS",
+    "ALL_SUPPORTED_DATASETS",
     "BLATM1BDataset",
     "Dataset",
     "GLAH06Dataset",
